Interact with GitHub API directly to add reviewers to a PR
The GitHub CLI has an issue where it will attempt to grab the
current set of reviewers when adding new reviewers. Since the
current set of reviewers will include the approvers team, it
will need access to see that team, which it does not have. This
causes the call to add reviewers to fail.

Instead we manually call the API, bypassing the need for this
permission.